Rename ModelsStore to Store
