+++ src/combineReducers.ts
@@ -1,4 +1,4 @@
-import { Action, Reducer, Reducers, State } from '../typings'
+import { State, Action, Reducer, Reducers } from "../typings/store"
 
 export function combineReducers<R extends Reducers>(reducers: R): Reducer<State<R>> {
   return (state: State<R> = {} as State<R>, action: Action) => {
