+++ src/compat.ts
@@ -1,5 +1,5 @@
 import { Store as ReduxStore, Reducer as ReduxReducer, AnyAction, Middleware, Observable } from 'redux'
-import { Store, ActionEvent, Reducer } from '../typings'
+import { Store, ActionEvent, Reducer } from '../typings/store'
 import { dispatchEvent, stateEvent } from './const'
 
 // compatibility wrapper to make store provide the Redux API
