+++ src/createModel.ts
@@ -1,3 +1,3 @@
-import { Model } from '../typings'
+import { Model } from '../typings/model'
 
 export const createModel = <S>(model: Model<S>) => model
