+++ typings/index.d.ts
@@ -3,7 +3,7 @@ export * from './connect'
 export * from './devtools'
 export * from './model'
 export * from './models'
-export * from './modelStore'
+export { Plugin, Plugins, Config, createStore, StoreState, StoreDispatch} from './modelStore'
 export * from './persist'
 export * from './routing'
 export * from './store'
