+++ typings/model.d.ts
@@ -12,8 +12,9 @@ interface EffectFns {
   [key: string]: EffectFn
 }
 
-// TODO: for consistency with state, rename dispatch to getDispatch
 export interface ModelStore<D = any, S = any> {
+  // TODO: for consistency with state, rename dispatch to getDispatch (?)
+
   // dispatch has to be a function that returns the Dispatch for the store
   // otherwise it creates a circular reference when added to the models' effects
   dispatch: () => D
