+++ typings/modelStore.d.ts
@@ -1,6 +1,6 @@
 import { Model } from './model'
 import { Models, ModelsDispatch, ModelsState } from './models'
-import { Store, Dispatch } from './store'
+import { Store as BaseStore, Dispatch } from './store'
 
 export interface Plugin<M extends Model = Model> {
   // if the plugin adds any state to the store, it can define it's own model
@@ -10,11 +10,11 @@ export interface Plugin<M extends Model = Model> {
   // NOTE: this is *NOT* the same model as above - this is called for each
   // model in the store as part of the setup, to give a plugin chance to do
   // whatever it needs to do based on each one.
-  onModel?<M extends Model>(store: ModelsStore, name: string, model: M): void
+  onModel?<M extends Model>(store: Store, name: string, model: M): void
 
   // TODO: thinking about it, why do we need 'onModel'? Can't we just handle
   // the store as a whole and the plugin can iterate the models if it wants?
-  onStore?(store: ModelsStore): void
+  onStore?(store: Store): void
 }
 
 export interface Plugins {
@@ -45,11 +45,11 @@ export interface Config {
   state?: any
 }
 
-export interface ModelsStore<M extends Models = Models> extends Store<ModelsState<M>> {
+interface Store<M extends Models = Models> extends BaseStore<ModelsState<M>> {
   dispatch: ModelsDispatch<M> & Dispatch
 }
 
-export declare function createStore<C extends Config>(config: C): ModelsStore<ConfigModels<C>>
+export declare function createStore<C extends Config>(config: C): Store<ConfigModels<C>>
 
 export type StoreState<C extends Config> = ModelsState<ConfigModels<C>>
